+++ lib/c/bitproto.c
@@ -1,4 +1,4 @@
-// Copyright (c) 2021, hit9. https://github.com/hit9/bitproto
+// Copyright (c) 2021~2023, hit9. https://github.com/hit9/bitproto
 // Encoding library for bitproto in C language.
 
 #include "bitproto.h"
@@ -6,6 +6,26 @@
 ///////////////////
 // Implementations
 ///////////////////
+
+// BpMin returns the smaller one int.
+static inline int BpMin(int a, int b) { return (a < b) ? a : b; }
+
+// BpMinTriple returns the smaller one of given three integers.
+static inline int BpMinTriple(int a, int b, int c) {
+    return (a < b) ? ((a < c) ? a : c) : ((b < c) ? b : c);
+}
+
+// BpIsNbitsStandard returns true if given nbits is one of 8/16/32/64.
+static inline bool BpIsNbitsStandard(int nbits) {
+    return nbits == 8 || nbits == 16 || nbits == 32 || nbits == 64;
+}
+
+// BpIsBaseNumberType returns true if given bp type flag is one of
+// byte/uint/enum/int. These types are all basic integer types in C.
+static inline bool BpIsBaseIntegerType(int flag) {
+    return flag == BP_TYPE_BYTE || flag == BP_TYPE_UINT ||
+           flag == BP_TYPE_ENUM || flag == BP_TYPE_INT;
+}
 
 // BpEndecodeMessage process given message at data with provided message
 // descriptor. It iterates all message fields to process.
@@ -105,29 +125,61 @@ void BpEndecodeArray(struct BpArrayDescriptor *descriptor,
         }
     }
 
+    int cap = descriptor->cap;
     int element_nbits = descriptor->element_type.nbits;
     int element_size = descriptor->element_type.size;
+    struct BpType *element_type = &(descriptor->element_type);
+
+    int flag = element_type->flag;
+    // The type flag behind if the element_type is an alias.
+    // For not alias type, to_flag will be 0.
+    int to_flag = element_type->to_flag;
+
     unsigned char *data_ptr = (unsigned char *)data;
 
-    // Process array elements.
-    for (int k = 0; k < descriptor->cap; k++) {
-        // Lookup the address of this element's data.
-        void *element_data = (void *)(data_ptr + k * element_size);
+    if (BpIsNbitsStandard(element_nbits) &&
+        (BpIsBaseIntegerType(flag) || BpIsBaseIntegerType(to_flag))) {
+        // Performance improvement for C arrays of integers (byte/uint/int):
+        // Since arrays in C are contiguous in memory layout, so we call
+        // BpCopyBufferBits only once instead of calling it one by one
+        // element, this makes huge performance improvement. The reason is that
+        // batch-copying technique works, the bits are copying every 4 bytes
+        // now. It's really a great optimization for arrays of complete integer
+        // types one of byte/uint8/uint16/uint32/uint64/int8/int16/int32/int64.
+        // andd enums of these uints, and alias to these types.
 
-        switch (descriptor->element_type.flag) {
-            case BP_TYPE_BOOL:
-            case BP_TYPE_UINT:
-            case BP_TYPE_BYTE:
-            case BP_TYPE_ENUM:
-                BpEndecodeBaseType(element_nbits, ctx, element_data);
-                break;
-            case BP_TYPE_INT:
-                BpEndecodeInt(element_size, element_nbits, ctx, element_data);
-                break;
-            case BP_TYPE_ALIAS:
-            case BP_TYPE_MESSAGE:
-                descriptor->element_type.processor(element_data, ctx);
-                break;
+        BpEndecodeBaseType(element_nbits * cap, ctx, data_ptr);
+
+        if (flag == BP_TYPE_INT || to_flag == BP_TYPE_INT) {
+            // We should handle the signed integer's sign after bits copying.
+            for (int k = 0; k < cap; k++) {
+                BpHandleIntSignAfterEndecode(element_size, element_nbits, ctx,
+                                             data_ptr);
+                data_ptr += element_size;
+            }
+        }
+
+    } else {
+        // Process array elements one by one.
+
+        for (int k = 0; k < cap; k++) {
+            switch (flag) {
+                case BP_TYPE_BOOL:
+                case BP_TYPE_UINT:
+                case BP_TYPE_BYTE:
+                case BP_TYPE_ENUM:
+                    BpEndecodeBaseType(element_nbits, ctx, data_ptr);
+                    break;
+                case BP_TYPE_INT:
+                    BpEndecodeInt(element_size, element_nbits, ctx, data_ptr);
+                    break;
+                case BP_TYPE_ALIAS:
+                case BP_TYPE_MESSAGE:
+                    descriptor->element_type.processor(data_ptr, ctx);
+                    break;
+            }
+
+            data_ptr += element_size;
         }
     }
 
@@ -140,107 +192,98 @@ void BpEndecodeArray(struct BpArrayDescriptor *descriptor,
     }
 }
 
-// BpCopyBits copy number of c bits from given char src to given char dst.
-// Returns the new value of dst.
-// The argument src_bit_idx is the index in the single byte src to start
-// coping. The argument dst_bit_idx is the index in the single byte dst to
-// start coping.
-unsigned char BpCopyBits(unsigned char dst, unsigned char src, int dst_bit_idx,
-                         int src_bit_idx, int c) {
-    // The number of bits to shift char src.
-    // If given src_bit_idx is smaller than dst_bit_idx, performs a shift
-    // left. Else shifts right.
-    int shift = src_bit_idx - dst_bit_idx;
-    // Mask to clear char src's right remaining bits all to 0.
-    // The following expression gives a number like 0b11110000.
-    unsigned char mask = (1 << (dst_bit_idx + c)) - (1 << dst_bit_idx);
-    // Result of a new dst byte.
-    // dst | (src >> shift) & mask
-    return dst | (BpSmartShift(src, shift) & mask);
-}
-
-// BpCopyBufferBits copy number of nbits from source buffer src to destination
-// buffer dst.
-// The argument n is the total number of bits to copy.
-// The argument src_bit_idx is the index to start coping on buffer src.
-// The argument dst_bit_idx is the index to start coping on buffer dst.
-void BpCopyBufferBits(int n, unsigned char *dst, unsigned char *src,
-                      int dst_bit_idx, int src_bit_idx) {
+// BpCopyBufferBits copy number of nbits from source buffer src to
+// destination buffer dst. The argument n is the total number of bits to
+// copy. The argument si is the index to start coping on buffer src.
+// The argument di is the index to start coping on buffer dst.
+// Note that this function DOESN'T suppose current processing type is a base
+// type, that's to say, we can process n bits copying where n could be a large
+// number than 64.
+void BpCopyBufferBits(int n, unsigned char *dst, unsigned char *src, int di,
+                      int si) {
     // n is the number of bits remaining to process.
-    while (n > 0) {
-        // Byte index in destination buffer.
-        int dst_byte_idx = dst_bit_idx >> 3;
-        // Byte index in source buffer.
-        int src_byte_idx = src_bit_idx >> 3;
+    while (n) {
+        // Byte index in buffer is (idx >> 3)
+        // where `>> 3` is faster than `/8`.
+        dst += (di >> 3);
+        src += (si >> 3);
 
-        // Remainder of destination bit index on 8.
-        int dst_bit_im = dst_bit_idx & 7;
-        // Remainder of source bit index on 8.
-        int src_bit_im = src_bit_idx & 7;
+        // Bit position inside current buffer byte.
+        // where `& 7` faster than `%8`
+        di &= 7;
+        si &= 7;
 
-        // Pointer to the destination to write this iteration.
-        unsigned char *dst_ptr = dst + dst_byte_idx;
-        // Pointer to the destination to read this iteration.
-        unsigned char *src_ptr = src + src_byte_idx;
-
-        // Number of bits to shift if goes into the optimized batch writing
-        // branch.
-        int shift = src_bit_im;
-        // The capacity of bits for batch writing.
-        int bits = n + shift;
-
-        // Number of bits to copy this iteration.
+        // c is the number of bits to copy in this iteration.
         int c = 0;
 
-        if (dst_bit_im == 0 && src_bit_im == 0 && n >= 8) {
-            // Assign directly if already aligned to 0.
-            dst_ptr[0] = src_ptr[0];
-            c = 8;
-        } else if (dst_bit_im != 0 || bits < 8) {
-            // Copy on single byte.
+        if (di == 0) {
+            // When di == 0, we can directly ASSGIN values from
+            // shifted src byte.
+
+            // Number of bits to process during batch copy.
+            int bits = n + si;
+
+            if (bits >= 32) {
+                // Copy as an uint32 integer.
+                // This way, performance faster x2 than bits copy approach.
+                ((uint32_t *)dst)[0] = ((uint32_t *)(src))[0] >> si;
+                c = 32 - si;
+            } else if (bits >= 16) {
+                // Copy as an uint16 integer.
+                ((uint16_t *)dst)[0] = ((uint16_t *)(src))[0] >> si;
+                c = 16 - si;
+            } else if (bits >= 8) {
+                // Copy as an unsigned char.
+                dst[0] = (src[0] >> si) & 0xff;
+                c = 8 - si;
+            } else {
+                // When bits < 8 and di == 0
+                // Copy partial bits inside a byte.
+                // For the original statement:
+                // c = BpMinTriple(8 - di, 8 - si, n);
+                // since di is 0 and bits <8, then 8-di is 8
+                // and n <8 , the 8-di won't be the smallest, we
+                // just pick function BpMin over BpMinTriple for the little
+                // little performance improvement.
+                c = BpMin(8 - si, n);
+                // Also, when di is 0, special case of next case.
+                dst[0] |= ((src[0] >> si) & ~(0xff << c));
+            }
+        } else {
+            // When di != 0, we have to copy partial bits inside a
+            // single byte.
+            // But, after some rounds of this case, di would goes to 0,
+            // for large sized types.
 
             // Number of bits to copy.
-            // 8-dst_bit_im ensures the destination space is enough.
-            // 8-src_bit_im ensures the source space is enough.
+            // 8-di ensures the destination space is enough.
+            // 8-si ensures the source space is enough.
             // nbits ensures the total bits remaining enough.
-            c = BpMinTriple(8 - dst_bit_im, 8 - src_bit_im, n);
-            dst_ptr[0] =
-                BpCopyBits(dst_ptr[0], src_ptr[0], dst_bit_im, src_bit_im, c);
+            c = BpMinTriple(8 - di, 8 - si, n);
 
-        } else if (bits >= 32) {
-            // Copy as an uint32 integer.
-            // This way, performance faster x2 than BpCopyBits iteration.
-            uint32_t v32 = (*((uint32_t *)(src_ptr))) >> shift;
-            unsigned char *ptr = (unsigned char *)(&v32);
-            dst_ptr[0] = ptr[0];
-            dst_ptr[1] = ptr[1];
-            dst_ptr[2] = ptr[2];
-            dst_ptr[3] = ptr[3];
-
-            c = 32 - shift;
-        } else if (bits >= 16) {
-            // Copy as an uint16 integer.
-            uint16_t v16 = (*(uint16_t *)(src_ptr)) >> shift;
-            unsigned char *ptr = (unsigned char *)(&v16);
-            dst_ptr[0] = ptr[0];
-            dst_ptr[1] = ptr[1];
-
-            c = 16 - shift;
-        } else if (bits >= 8) {
-            // Copy as an unsigned char.
-            dst_ptr[0] = (src_ptr[0] >> shift) & 0xff;
-
-            c = 8 - shift;
+            // Explaination:
+            // src >> si << di  Margins byte src with byte dst at position
+            // di. this also clears the right bits up to position si.
+            // ~(0xff << di << c) Gives a mask to clear higher not-need bits
+            // all to 0, e.g. 00001111 on di=4, c=4; Finally: dst |= src to
+            // copy bits.
+            // The ch is the first byte at pointer src. If this byte is Zero,
+            // then there's no need to copy anything, just count c. The
+            // benchmark on stm32 seems performance is improved by 2us by adding
+            // an if statement that skip the Zero byte.
+            unsigned char ch = src[0];
+            if (ch) dst[0] |= ((ch >> si << di) & ~(0xff << di << c));
         }
 
         // Maintain in(de)crements.
         n -= c;
-        dst_bit_idx += c;
-        src_bit_idx += c;
+        di += c;
+        si += c;
     }
 }
 
 // BpEndecodeBaseType process given base type at given data.
+// This function guarantees to work geven a nbits > 64 is passed in.
 void BpEndecodeBaseType(int nbits, struct BpProcessorContext *ctx, void *data) {
     if (ctx->is_encode) {
         BpCopyBufferBits(nbits, ctx->s, (unsigned char *)data, ctx->i, 0);
@@ -250,15 +293,14 @@ void BpEndecodeBaseType(int nbits, struct BpProcessorContext *ctx, void *data) {
     ctx->i += nbits;
 }
 
-// BpEndecodeInt process signed integer at given data.
-// The most left bit (Nth bit for int{N}) of a signed integer indicates the
-// sign. For example 00000101 is a negative integer for a int3, but a positive
-// integer for a int4.
-void BpEndecodeInt(int size, int nbits, struct BpProcessorContext *ctx,
-                   void *data) {
-    // Copy bits without concern about sign bit.
-    BpEndecodeBaseType(nbits, ctx, data);
-
+// BpHandleIntSignAfterEndecode processes signed integer at given data after
+// this integer is endecode. The most left bit (Nth bit for int{N}) of a signed
+// integer indicates the sign. For example 00000101 is a negative integer for a
+// int3, but a positive integer for a int4. Where nbits is the number of bits
+// for this bitproto signed integer, for int{N}, its the N, the argument size is
+// the number of bytes in C language.
+void BpHandleIntSignAfterEndecode(int size, int nbits,
+                                  struct BpProcessorContext *ctx, void *data) {
     // Signed integer's sign bit processing is only about decoding.
     if (ctx->is_encode) return;
 
@@ -266,8 +308,8 @@ void BpEndecodeInt(int size, int nbits, struct BpProcessorContext *ctx,
     // most-left bit position. There's no additional actions should be done.
     if (nbits == 8 || nbits == 16 || nbits == 32 || nbits == 64) return;
 
-    // Number of bits occupied in C intXX_t types.
-    int n = 8 * size;
+    // Number of bits occupied in C intXX_t types, aka: n = size * 8
+    int n = size << 3;
 
     switch (n) {
         // Suppose pointer data points to a int24 value V:
@@ -278,9 +320,9 @@ void BpEndecodeInt(int size, int nbits, struct BpProcessorContext *ctx,
         //    Recover the original integer is: V | mask.
         //
         // Why not use V << 8 >> 8 solution here? This depends on arithmetic
-        // shifting. Which propagates the sign bit on the left vacant bits when
-        // doing a right shift. But C standard points that this behavior is
-        // implementation-defined.
+        // shifting. Which propagates the sign bit on the left vacant bits
+        // when doing a right shift. But C standard points that this
+        // behavior is implementation-defined.
         case 8:  // int8_t
             if (((*(int8_t *)data) >> (nbits - 1)) & 1) {
                 *(int8_t *)data |= ~((1 << nbits) - 1);
@@ -302,6 +344,15 @@ void BpEndecodeInt(int size, int nbits, struct BpProcessorContext *ctx,
             }
             break;
     }
+}
+
+// BpEndecodeInt process a single signed integer at given data.
+void BpEndecodeInt(int size, int nbits, struct BpProcessorContext *ctx,
+                   void *data) {
+    // Copy bits without concern about sign bit.
+    BpEndecodeBaseType(nbits, ctx, data);
+    // Handle the signed bit.
+    BpHandleIntSignAfterEndecode(size, nbits, ctx, data);
 }
 
 // BpEncodeArrayExtensibleAhead encode the array capacity as the ahead flag
@@ -340,20 +391,6 @@ uint16_t BpDecodeMessageExtensibleAhead(struct BpMessageDescriptor *descriptor,
     uint16_t data = 0;
     BpEndecodeBaseType(16, ctx, (void *)&data);
     return data;
-}
-
-// BpMinTriple returns the smaller one of given three integers.
-int BpMinTriple(int a, int b, int c) {
-    return (a < b) ? ((a < c) ? a : c) : ((b < c) ? b : c);
-}
-
-// BpSmartShift shifts given unsigned char n by k.
-// If k is larger than 0, performs a right shift, otherwise left.
-unsigned char BpSmartShift(unsigned char n, int k) {
-    // Quote: If the value of the right operand is negative or is greater
-    // than or equal to the width of the promoted left operand, the behavior
-    // is undefined.
-    return (k > 0) ? (n >> k) : ((k == 0) ? n : (n << (0 - k)));
 }
 
 // BpJsonFormatString is a simple wrapper on sprintf that accepts